+++ src/main/java/com/example/infrastructure/persistence/entity/OrderPo.java
@@ -10,6 +10,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -19,24 +20,24 @@ import java.time.LocalDateTime;
 @Entity
 @Table(name = "order")
 public class OrderPo {
-  @Id
-  @GeneratedValue(generator = "system-uuid")
-  @GenericGenerator(name = "system-uuid", strategy = "uuid")
-  private String id;
+    @Id
+    @GeneratedValue(generator = "system-uuid")
+    @GenericGenerator(name = "system-uuid", strategy = "uuid")
+    private String id;
 
-  private String customerId;
+    private String customerId;
 
-  private Double total;
+    private BigDecimal total;
 
-  @Column(columnDefinition = "TEXT")
-  private String products;
+    @Column(columnDefinition = "TEXT")
+    private String products;
 
-  private OrderStatus status;
+    private OrderStatus status;
 
-  @CreatedDate
-  private LocalDateTime createTime;
+    @CreatedDate
+    private LocalDateTime createTime;
 
-  @LastModifiedDate
-  private LocalDateTime updateTime;
+    @LastModifiedDate
+    private LocalDateTime updateTime;
 
 }
